+++ CHANGELOG.md
@@ -12,6 +12,8 @@ All notable changes to this project will be documented in this file.
 - Description: RBY1 Mobile Manipulatior (URDF) (thanks to @uynitsuj)
 - Description: YAM (MJCF) (thanks to @uynitsuj)
 - Description: YAM (URDF) (thanks to @uynitsuj)
+- Description: Unitree H1_2 (URDF) (thanks to @TonyZYT2000)
+- Description: Unitree H1_2 (MJCF) (thanks to @TonyZYT2000)
 
 ## [1.19.0] - 2025-07-03